compiler/natives/src/reflect: Change makechan size parameter type from uint64 to int.
It has changed upstream.

Related to https://github.com/golang/go/commit/8a6e51aede746d1f7ffec5982c153bf3cdabce4f.

Fixes:

	reflect/value.go:2119:31: cannot use buffer (variable of type int) as uint64 value in argument to makechan

+++ compiler/natives/src/reflect/reflect.go
@@ -478,7 +478,7 @@ func loadScalar(p unsafe.Pointer, n uintptr) uintptr {
 	return js.InternalObject(p).Call("$get").Unsafe()
 }
 
-func makechan(typ *rtype, size uint64) (ch unsafe.Pointer) {
+func makechan(typ *rtype, size int) (ch unsafe.Pointer) {
 	ctyp := (*chanType)(unsafe.Pointer(typ))
 	return unsafe.Pointer(js.Global.Get("$Chan").New(jsType(ctyp.elem), size).Unsafe())
 }
